Use xdg-open as the default browser command on Linux if available.
'xdg-open' is a cross-desktop-environment tool that automatically
detects the appropriate web browser; use it if it is available,
preferred over the desktop environment specific tools.

This fixes the browser detection on XFCE and Gnome 3 (which no longer
includes the gnome-open tool).

This change only fixes the code for python >=2.5, I have no way of
testing older versions.