Don't use native file dialog with multiselect on (PICARD-74)
On OS X, the native file dialog is used by default when a new QFileDialog
is created. It can't be configured to select multiple directories, which
breaks the toolbar_multiselect option. If this is enabled, disable the
native file dialog on OS X.

This option is actually rather misleading, because on Windows 7 the
native file dialog is also used by default, and it can select multiple
directories just fine. However, windows only seems to use the native
dialog when the static helper functions are called, not when a
custom QFileDialog instance is created.