Store lyrics language in tags instead of text representation language (PICARD-242)
This changes the meaning of the "language" tag. It is now no longer
derived from the text representation language of the release, but from
the lyrics language of the linked work(s).

The text representation language of the release is still available (e.g.
for scripting/plugins) as %_releaselanguage%. To get the old behavior
back, one must insert this line at the beginning of the scripting field:

    $set(language,%_releaselanguage%)

http://tickets.musicbrainz.org/browse/PICARD-242