Try to solve UI freezes
Calling setCurrentIndex from currentChanged causes an infinite loop, breaking
the UI. But without that line, there are other very strange effects: see
PICARD-248. The only thing that currentChanged was added for was to make sure
the next item is selected after something is removed, so that keyboard
navigation is easier. This implements a new workaround for that without using
currentChanged.