Enable Informations dialog for albums, used to display metadata images.
It is using existing InfoDialog code, moved to a common class.
For now, first tab is just hidden as there is nothing to display in.
In Artwork tab, all metadata images associated with the album are shown.
Slight change in behavior is that pressing Informations button while an album is
selected was showing 1st track information, now it should album info.