CAA downloader: fix CAA download of 'unknown' type images
When no type is set for an image, CAA is returning an empty list for types.
But Picard user has no way to specify he wants to also download images
of unknown type using the option 'Download only images of the following types'.

I added a special type named 'unknown' and modified _caa_json_downloaded() to
handle this case.