addrelease: Try harder to seed the RE properly
Disc numbers <= 0 were previously changed to values lower than 0 which
caused the RE to crash. Fix this by using the lowest disc number of the
cluster (which is somewhat sorted) to figure out the difference between
that disc number and 0 and use that number to change all disc number to
be 0-based while preserving their order.