PICARD-296: Cover file is saved repeatedly
If overwrite was enabled, the else suite was executed which saved the
cover file once for each track of an album even if there was only one
file. Check if the filesize is the same as that of the file we're trying
to save - if it is, don't save the file again.