Remove sorting for displayed multi-value tags and provide a sort tags plugin
Picard stores tags unsorted but displays them sorted and this is
inconsistent.

This commit removes display sorting of multi-value tags from Picard
improves consistency by displaying what will be / is actually saved.

This commit also provides a plug-in to allow user to sort tags both
displayed and stored in the file.