Make track lengths within 1s of each other match
Due to minor timing differences when ripping a CD, it is common for
tracks to differ by up to 1s in length.

This commit enhances metadatabox to show lengths +/- 1s as matching.

Note: Means of converting mmm:ss to seconds is not elegant - anyone with
a better solution please say.