Force use of network cache if enabled
If user has set the Force Cache option, then if it is a GET request we
check to see if it is cached, and set cache control options accordingly.

If we are setting QNetworkRequest.AlwaysCache, then we will not be
sending a request to a web site, so we will avoid resetting the last
request time.