Add support for loading coverart from local files.
I have coverart that I'm unlikely to ever upload to e.g. the Cover Art
Archive.  This commit teaches picard to load it and add it to files.

Loading local cover art, and the name of the cover art file, are
managed by config options.

When a file is loaded, picard now searches for a local cover art file
in the same directory.

If a file is found it is added to the file's metadata.  When the file
becomes associated with a track, the file's images are added to the
track's images.  If the track's album's metadata has no associated
images, the file's image is also added to the album's metadata.