Add option to limit browser integration to localhost.
This will make Picard's browser integration to listen on localhost only, as listening on the public interface is a possible security and privacy concern. There is still an option to enable listening on the external interface to allow for special requirements such as using the Picard Barcode Scanner android app.