PICARD-123: actually use default drive on Mac
This doesn't implement an automatic list of available drives on Mac,
but adds the libdiscid default drive to the list of drive candidates.

This should make disc lookups available on Mac without specifying
the drive (as /dev/rdisk*) in the settings.

Note that setting things like "1" or "2" as drives should also work
with a recent libdiscid. (>= 0.6.0).