Set preferred releases to "Digital Media" and "[Worldwide]" by default.
If two releases with same tracklist exists, we want to take digital media over CD (ie),
because most users are tagging digital files from unknown source and don't care
which exact release it is until tracklist is matching.

Worldwide is preferred by default, because if two releases exist with same tracklist
it is more likely it is the worldwide one's than a country-specific one.