Copy language from release if not in track
Track languages are taken from the related work, however if the track
does not have a related work then language is not set even if it is set
in the Release.

This fix copies the release language to the track if there is no
language received from a related work.