Move work to TOAL tag.
TOAL is defined in
>  TOAL - The 'Original album/movie/show title' frame is intended for the title of the original recording (or source of sound), if for example the music in the file should be a cover of a previously released song.

Whilst this is not explicitly the correct definition of Work Title, in practical terms it is as the Work typically is named the same as the original recording. In particular this will provide proper id3v23/4 support for classical music where the Track and Work titles are oftent very different.

Provide translation for anyone who has used daily builds with TXXX:Work.