Improve script check for number of args.
Remove the need for specifying check_argcount=False if function uses
*args.

expected_args is now a 2-value tuple with lower and upper range of
argument counts.

Also, looks like a bug in calc if there were defaults.

Commented print string left in so correct range can be verified if
needed.