Move ID3Metadata functionality to new_metadata property
The problem with ID3Metadata is that it changes the behavior of
Metadata methods in undesirable ways (see PR #273).

This adds a new_metadata property to File classes which returns a copy
with the values modified as they'd be displayed/saved. This should more
composable and less error-prone than overriding Metadata methods across
the board.