Added a menu option to add new files to tagger
A file is considered 'new' if it has been modified in the last five
hours. This setting can be stored in the config later.

Only the starting directory is scanned for new files currently. A
context menu option can be added to the treeview to add new files from a
folder.