Load new files while adding a directory
The file menu option has been removed and has been replaced with a
combobox in the "Add Directory" dialog.

The combobox is uneditable for now and just contains some pre-defined
periods as there didn't seem an easy way to take input from user.

The dialog looks a bit odd with a button 'missing' -
http://i.imgur.com/QOmhzr2.jpg