Ignore 404 errors for release group json
We currently have no easy mean to know if there is cover art for the release group as we do for the release.

Not a perfect solution, ideas and patches are welcome.
We should not query the json file at all if we know there is no cover art for this release group.