Refactored loading of plugins list
Previously, the data was first written to a file and then read again
by different options pages - which was just stupid.

Now, the data is loaded as json right when its download.
All other pages can then just use it.

A statusbar message is set if an error occurs in loading the list.