Fix annoying little bugette in APEv2
Code was tweaked to handle invalid cover art in a file saved by musicbee
(marked as TEXT not BINARY) but this introduced a minor bug if there was
a normal tag starting with "cover art" like "Cover Art URL".

Since cover art actually starts with "cover art (" we can use this to
address this.