Quickfix for upstream Mutagen compatibility
There have been some changes in Mutagen upstream which makes it
incompatible with Picard. For reference, SHAs of commits in Mutagen
upstream (https://github.com/quodlibet/mutagen) that currently break Picard:

1. 2cd8cfdaec88bf896afd46cd03b209d688d3740a
2. e6e3fbe34f5946c92be8fb9266e9d27b46d8a700

This patch takes care of things for now, but is not a permanent
solution. Also as Mutagen has dropped support for Python 2.6, but Picard
is continuing it, there may be more conflicts in future.