Try capturing focus in event rather than out
The accept button needs to be disabled to allow searching on pressing
return. If not disabled, the dialog closes on return. In commit a44c375,
focus out event of the table was used to disable the load button. This
doesn't works as the button will be disabled when user moves the mouse
towards the button (focus moves out of the table).
To avoid this, capture focus in event of `search_edit` widget.