PICARD-851: Correctly update parent items on save
When multiple files were linked to the same track and the changes to
the items were saved, the parent album icon didn't update to reflect
the change.

This was because of a buggy item.save() function which led to an update
call to the update() method of files. The files on update did not recu-
rsively call their parent items to be updated too.