PICARD-855: Adds errorhandler for missing meta-data
When a track with a non valid total-tracks has multiple release matches
on Picard, picard tries to apply a linear combination  of weights to
detect the best match. In the process it tried to typecast totaltracks
into and int.

This causes a ValueError when totaltracks is not castable into an int.