PICARD-897: Rename "Submit" button to "Submit AcoustIDs".
The submit button doesn't actually submit any information to
MusicBrainz, but instead submits acoustic fingerprint information to
AcoustID.org. The title of this button comes up over and over as a point
of confusion for new users of Picard, so renaming it to be more precise
will hopefully alleviate this point of confusion.