Bump minimum Python version to 2.7.9 for Picard 1.4.
Since the various MetaBrainz projects are increasingly moving to
HTTPS-mostly (and hopefully -only before too long!), it would be nice to
know that Picard 1.4 won't break because of this anytime soon. Since
several of the MetaBrainz sites share an IP address they're relying on
SNI for sending the proper certificate. SNI support wasn't added in
Python 2.7 before the .9 release, so bumping minimum Python version to
this specific one, so that we can be sure that SNI is supported.

See also:
https://chatlogs.metabrainz.org/brainzbot/metabrainz/msg/3800847/