PICARD-951: Always use HTTPS for musicbrainz.org.
MusicBrainz.org is eventually going to go all HTTPS-only, though the
webservice for now can still be reached via HTTP. To help the move to
HTTPS, this commit makes URLs built via `build_qurl()` always use HTTPS
when going to a MusicBrainz server (as defined by the
`MUSICBRAINZ_SERVERS` constant).

This also means that the check in the `build_qurl()` function for
whether a given page/URL is/requires an authentication for MusicBrainz
is now redundant, as we're always using HTTPS on MusicBrainz, regardless
of login/authentication status.