PICARD-953: Show album with extra tracks unmatched
Resolves [PICARD-953](https://tickets.metabrainz.org/browse/PICARD-953).

If an album has unmatched files ii is shown as incomplete even if all album tracks are matched to one file. Although the terminology would suggest that such an album is "complete", in practical terms the album is not perfect because in all likelihood the user needs to select a different release with more tracks.