PICARD-969: Queue search dialog xmlws to run first
If you are loading a lot of albums and have a long queue of releases
being looked up, then the xmlws request for Search for similar albums...
is queued behind them.

This commit runs these request as priority=True important=True which
means that they will run as the very next xmlws.