avoid multiple value TXXX frames (PICARD-990)
Multiple values in TXXX frames are supported by Mutagen, but violate the official ID3 specification. This would allow all compliant ID3 implementations to read these tags, as opposed to just the ones violating the spec in this way. I believe this change will work, but I wasn't able to build picard to test it. Whether it's the best approach is another question. Related to https://tickets.metabrainz.org/browse/PICARD-990