Small refactorization of TrackInfoDialog
In order to fix tracks without files not having the info button
available, I changed it to allow 0, 1 or multiple files.
If the track doesn't have any linked file, only the cover art is shown
(if available). If the track has n files, the info tab shows information
from the all the files.