Make front cover glow when there are no common images among tracks
If all tracks/files have a common set of images, show the stack
of cover art images as usual. If any track/file has a different
set of images, the front image is painted with a darkgoldenrod glow
(the same color used for the MetadataBox for tags with different values).

Also, fixed the draw of the stack so the bottom cover is drawn
correctly.