PICARD-1010: Fix display/hide logic for Coverart box
Earlier when the display for Cover art box was toggled, it only
hid the cover art box on trigger. On selection update a
super().show() was called which again caused the cover art box to be
visible. This fixes the flawed display logic due to PR#643.

See: https://tickets.metabrainz.org/browse/PICARD-1010