Use self.has_common_images to know if set_data should be executed or not
The images used by the Thumbnail are kept in a "poor-man-cache" to know
if set_data should be run or not, but has_common_images wasn't used.
This fixes the situation where an album with only two files in
different tracks, one with two original images and another with none,
would keep a wrong original cover art when selecting the album and the
track with the two original images alternatively. The reason is that
the images to show would be the same, but for the Track, has_common_images
would be True, while for the Album, it would be False (so for the Track
it would be a normal stack but for the album it would have the yellow glow)