Fix comparisons of list of images sorting them in coverartbox and infodialog
When the list of new images and the list of original images are compared
to select if there are changes, it's important that they are sorted,
since when an info dialog is opened, it sorts the images by image type.
This breaks all future comparisons of that list of images with others,
so this sorts the images to fix this issue.
This fixes PICARD-1001