Allow to view info also when it only has orig_metadata.images
An album can have no metadata.images but have orig_metadata.images.
In that case, we also want to display the info dialog, to show
the existing images.
Also, removed self.metadata, since for an album, that's only true when
self.metadata.images is true, which is tested right afterwards.