Fix info dialog image diff corner cases
Make InfoDialog store locally in the constructor the images it'll use as
new images and as existing images. This allows to swap them in case we
want to display the existing artwork of a file whose release doesn't
define any artwork (that is, a track with 1 file and which has
orig_metadata.images but no metadata.images). Also, it'll allow
for easier refactoring in the future.
Also make CoverArtBox use item.can_view_info instead of duplicating
(wrong) the logic to decide if the "show more info" button should be
visible or not.