Add message "1 image not in all tracks" and fix corner case
When one track has one image and the rest don't have any. Show
"1 image not in all tracks" as album title.
Also, count images in orig_metadata.images when a release doesn't
have any, since CoverArtBox shows orig_metadata in such a case and
it would be strange to have a cover shown and "0 images" in the title.