Compare CoverArtThumbnails using the hash key of the current used pixmap
By also setting has_common_images to True when len(data)==1 (which doesn't
affect the shown pixmap) when calculing the key, this fixes the case
when an album has only 2 files, one with the same images as the release
and the other doesn't have any image at all. In that case, when
choosing if orig/new widgets have to be visible, the original
and new images were the same, but the "new" widget had
has_common_images == True and "orig" had has_common_images == False.
Since the second file actually doesn't have any orig image, we don't
want to show orig/new widgets when selecting the album.