Fix file state not changing to NORMAL when no image was provided
If the user disabled all cover art providers and tried to save a file
that already contained an image, the state would never change to normal
even after saving the file. This commit fixes this by only comparing
images in File.update if the metadata.images variable is not empty
(note that metadata.update only copies image list if it's not empty,
so this makes sense).