Remove images in keep_original_images if the track doesn't have a file
If the track doesn't have a file and the user selected to keep original
images, remove the downloaded images from metadata.images so when a
file is added to the track, the original image is kept. Also, this
is more consistent with respect to album images for albums where
some tracks don't have a file and "Keep original images" is selected.