Fix multi-directory load on Windows
On Windows on Qt5 it appears you need to avoid native dialog.

As far as I can tell, this was actually a bug under Qt4 (because native
dialog wasn't used anyway) and it should always have been specified.

I have only added win32 to the exception list. This may be required on
all platforms, but I am unable to test this.