Use secondary types in release-type algorithm
The old algorithm that determined the score of the release based on
the release type used only the primary release type and ignored the
secondary types. This change, averages the score for all types found,
both primary and secondary. This change also sets the score to
the value of 'Other' if no release type is found.