Use the right shebang in the main executable
On most platforms, the unversioned "python" executable still points to
Python 2.

As a result, the `env python` shebang will always give a Python 2
interpreter on those platforms.

But when installing Picard with `python3 setup.py install`, then a
Python 3 interpreter needs to be used to run the script.

This change sets the shebang for the installed script to the Python
executable used to run the setup.py script.