PICARD-1207: Fix handling paths with unicode characters on python3
When moving additional files, the call to encode_filename encodes the
path to the original filename converting the variable from str to bytes,
provoking the next exception when calling shutil, which only accepts
strings (the documentation don't specifically mention that for shutil.move,
but it does for copy, copyfile, copymode, etc.)

Traceback (most recent call last):
  File "./picard/util/thread.py", line 47, in run
    result = self.func()
  File "./picard/file.py", line 222, in _save_and_rename
    self._move_additional_files(old_filename, new_filename)
  File "./picard/file.py", line 435, in _move_additional_files
    shutil.move(old_file, new_file)
  File "/usr/x86_64-pc-linux-gnu/lib/python3.6/shutil.py", line 551, in move
    if _destinsrc(src, dst):
  File "/usr/x86_64-pc-linux-gnu/lib/python3.6/shutil.py", line 565, in _destinsrc
    if not src.endswith(os.path.sep):
TypeError: endswith first arg must be bytes or a tuple of bytes, not str

Also, other functions like os.path.dirname, os.path.isdir, os.stat,
os.utime ... allow to be passed a string, instead of the encoded path in
a bytes variable. So I removed the use of encode_filename for all cases
that either require or allow a string to be passed so we don't do
unnecessary processing neither.

I checked this fixes PICARD-1207.