PICARD-1369: Retry opening URLs
This is a workaround for https://bugs.python.org/issue31014. The first time
webbrowser.open is called, an exception is thrown when initializing the list of
browsers on the system. Ignoring that first exception leaves the webbrowser
module in a possibly-not-optimal state regarding the list of browsers it uses,
but it does successfully open URLs. Since there's no point in fixing this in
Picard itself, just continue with a not optimal browser choice.