Fix a crash when parsing an incomplete regular expresion being written
If you write $rsearch(text,) in the file naming script and then move the
cursor after the comma and start writing a regular expression like
[a-z], picard crashes as soon as the '[' character is entered, since it
tries to parse the regular expression and the re module raises an exception.

This patch catches the re exceptions for $rsearch and $rreplace so
picard doesn't crash.