Share data between artist nodes
This patch temporarily caches release artist nodes, and copies them to
the release group, track, and recording artist credits wherever the
artist IDs match. This allows Picard and external plugins to access
supplementary data (aliases, tags, genres, ratings) no matter where the
artist node was parsed. A similar change has been made to copy track
artist nodes to recordings.

It's true that for some data, like aliases in particular, the web
service already duplicates this information to every artist node.
However, I'd like to change this on the MBS side before we announce
that the web service is out of beta (and it already won't be true for
tags or genres). Having Picard work in either case would allow us to
make that change in the future.

Perhaps this patch also allows duplicate data to be GC'd faster, though
I didn't perform any profiling and that's not the aim here.