Better handling of missing libdiscid, catching OSError as per documentation
Picard cannot run without this patch and missing libdiscid and python-libdiscid.
It falls back on discid module, but when this one is trying to load libdiscid it fails with:
OSError: libdiscid.so.0: cannot open shared object file: No such file or directory

According to https://python-discid.readthedocs.io/en/latest/install/#dependencies

"The module discid cannot be imported without Libdiscid >= 0.2.2 installed.
If you want to use it as optional dependency, import the module only when
needed or catch the OSError when libdiscid is not found."

So catch OSError and define discid to None.
After that, Picard runs but emits a warning:
W: 10:54:37,425 picard/ui/mainwindow.create_actions:465: CDROM: discid library not found - Lookup CD functionality disabled