PICARD-1688: Submit AcoustID fingerprints in batches
Avoids issues with exceeding the body size limit for submission requests. Also this makes the entire submission process more robust, as the submission of many fingerprints can now succeed partially instead of failing in its entirety.

The upper limit is around 250 fingerprints, which is equivalent to roughly 1 MiB of data. But use a batch count of actually 240 fingerprints per submission to have some leeway for varying data.