PICARD-1750: Add AcoustID fingerprint from tags to AcoustIDManager
If a file has an existing AcoustID fingerprint in its tags and ignore_existing_acoustid_fingerprints is not set, then use this fingerprint to properly initialize the AcoustIDManager. This makes the existing fingerprint available for submission.

Previously such a file would not have its fingerprint used even after a Scan.