add album_save_post_processors to plug-in API.
I wanted to extend the M3U playlist plugin to automatically generate an .m3u8
when Albums get saved.

To be able to do so, I needed to add album_save_post_processors so that I could
get access to an album on save.

File save is done in a thread, so I needed to implement a way to wait for those
threads to finish by using the file_save_post_processors hook.