PICARD-1963: Avoid stack overflow when loading large amount of files
When loading a large list of files this could result in a stack overflow,
since tagger.move_files calls QCoreApplication.processEvents, which in
return might process an event to call tagger.move_files again, resulting
in a deep recursion.

Refactor the special case of moving a single file into a separate method.