Rework local cover art handling
Local cover art as a provider doesn't make sense - local cover art is a
property of a file, same as tag cover art. Move the handling of local
cover art to the File class.

TODO: Need to resolve how to configure load ordering of tag cover art
vs local cover art. Also, there needs to be a config ui written.

In the current state, the local_cover_regex setting is re-used, and a
new load_local_cover_art boolean is used to enable the feature.

This will probably fix many of the causes of PICARD-1001.