Send tport with Add cluster as release
When you Add cluster as release, the tport is lost on MBS and Green Tagger buttons disappear.

Thanks @phw for QCoreApplication.instance().browser_integration.port hint!

MBS does not accept tport as POST parameter (nknown field: tport).

It works when added in the action URL, even if it is ugly to mix GET query paramers and POST body parameters like this. :-/