Remove legacy genre tag translation
The translation mapping was originally added as part of the last.fm
plugin in commit 99dc3e54, which was later used as the basis of the
initial folksonomy genre implementation.

This change fixes a bug where a track tagged with the "electronic" and
"electronica" genres would end up with "electronic" twice.