fallback used, filename tests removed, should work fine
I've decided to remove the filename tests becaue they provide no useful value during the tests. Everything has to be mocked and so on... Debug-logging the pipe path should do, because during the tests we cannot be 100% sure that the specific folder will be writable, the env variable will be declared etc.