PICARD-2627: Ensure pipe listener threads closes on quit
It can happen that tagger.pipe_server hangs waiting getting data from the
pipe. If this happens when Picard is closing the pipe_server thread keeps
running in the background until it reveives data, which likely won't
happen. To fix this write a MESSAGE_TO_IGNORE to the pipe when closing.