Avoid a warning on Picard start-up about send_to_pipe
Before this patch, Picard is logging a warning before anything else that looks like:

`W: pipe.send_to_pipe:276: Couldn't send: '\x00'`

At this point, the pipe isn't fully functional, and this message is expected, as Picard sends a message to pipe to see if it's the owner or not (in `AbstractPipe.__init__()`).
The behavior is normal, the warning isn't.

To fix it, check if the pipe is marked as running before logging such warning.