Do not use a thread pool for read_from_pipe()
This method is already meant to run in a loop in a separate thread,
no need to spawn additional threads. Also this does not need to timeout
all the time. It is ok the thread waits until data becomes available
instead of looping all the time using CPU resources. Stopping the
pipe reader thread is done on application exit by a message.