Add a way to select extra debug output
Web service replies can be very long, and it makes debug output rather hard to read.

A debug option mechanism is now added, it is controlled by a new class
`DebugOpt`.

A command-line option `--debug-opts` allows selection of extra debug
output, for now the sole option is `ws_replies`.

A matching menu was added in View Debug Log dialog.

It is important to note those aren't saved in config file, as they are
meant to be disabled by default and only enabled when needed.