PICARD-2940: Fix prematurely releasing lock in `Metadata.__iter__`
Fixes https://tickets.metabrainz.org/projects/PICARD/issues/PICARD-2940

Previously the lock was being acquired and then released immediately when the iterator was returned. This meant that when the iterator was being iterated by the consumer, the lock was not held.

This change ensures that the lock is held until the iterator is exhausted or goes out of scope.