Sanitize HTTP header values in browser integration
This is a fix for potential HTTP Response Splitting, as the
Access-Control-Allow-Origin was set from incoming data.
In practice this was not a vulnerability, though. For one this data
was read from HTTP headers and cannot contain line breaks.
But also origins are validated against a known list of valid hostnames.

Fixes security code scanning alerts #2591 and #2592